# backend/src/main/java/com/urbanfix/dto/CategoryAnalyticsResponse.java
package com.urbanfix.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CategoryAnalyticsResponse {

    private String category;

    private long complaintCount;

}