# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:5050","description":"Generated server url"}],"paths":{"/api/complaints/{complaintId}":{"get":{"tags":["complaint-controller"],"operationId":"getComplaintById","parameters":[{"name":"complaintId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["complaint-controller"],"operationId":"updateComplaint","parameters":[{"name":"complaintId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateComplaintRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["complaint-controller"],"operationId":"deleteComplaint","parameters":[{"name":"complaintId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/api/complaints":{"get":{"tags":["complaint-controller"],"operationId":"getAllComplaints","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","IN_PROGRESS","RESOLVED","REJECTED"]}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["complaint-controller"],"operationId":"createComplaint","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"request":{"$ref":"#/components/schemas/CreateComplaintRequest"},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/complaints/{id}/status":{"patch":{"tags":["complaint-controller"],"operationId":"updateComplaintStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateComplaintStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/users/me":{"get":{"tags":["user-controller"],"summary":"Get the authenticated user's profile","operationId":"getCurrentUserProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/users/me/dashboard":{"get":{"tags":["user-controller"],"summary":"Get the authenticated user's dashboard statistics","operationId":"getCurrentUserDashboardStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDashboardStatsResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/test":{"get":{"tags":["test-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/complaints/search":{"get":{"tags":["complaint-controller"],"operationId":"searchComplaints","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/complaints/my":{"get":{"tags":["complaint-controller"],"summary":"Get the authenticated user's complaints","operationId":"getMyComplaints","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","IN_PROGRESS","RESOLVED","REJECTED"]}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageComplaintResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/complaints/dashboard":{"get":{"tags":["complaint-controller"],"operationId":"getDashboardStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardStatsResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth":{"get":{"tags":["auth-controller"],"operationId":"test_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/admin/dashboard":{"get":{"tags":["admin-controller"],"operationId":"getDashboardStatistics_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardStatsResponse"}}}}}}},"/api/admin/dashboard/monthly":{"get":{"tags":["admin-controller"],"operationId":"getMonthlyComplaintAnalytics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MonthlyComplaintAnalyticsResponse"}}}}}}}},"/api/admin/dashboard/categories":{"get":{"tags":["admin-controller"],"operationId":"getCategoryAnalytics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryAnalyticsResponse"}}}}}}}},"/api/admin/complaints":{"get":{"tags":["admin-controller"],"operationId":"getAllComplaints_1","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","IN_PROGRESS","RESOLVED","REJECTED"]}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0,"minimum":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10,"minimum":1}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"sortDirection","in":"query","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageComplaintResponse"}}}}}}}},"components":{"schemas":{"UpdateComplaintRequest":{"type":"object","properties":{"title":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"category":{"type":"string","minLength":1},"location":{"type":"string","minLength":1},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"imageUrl":{"type":"string"}},"required":["category","description","location","title"]},"ComplaintResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"category":{"type":"string"},"location":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"imageUrl":{"type":"string"},"status":{"type":"string","enum":["PENDING","IN_PROGRESS","RESOLVED","REJECTED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userName":{"type":"string"}}},"CreateComplaintRequest":{"type":"object","properties":{"title":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"category":{"type":"string","minLength":1},"location":{"type":"string","minLength":1},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"imageUrl":{"type":"string"}},"required":["category","description","latitude","location","longitude","title"]},"RegisterRequest":{"type":"object","properties":{"fullName":{"type":"string","minLength":1},"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","fullName","password"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"}}},"UpdateComplaintStatusRequest":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","IN_PROGRESS","RESOLVED","REJECTED"]}},"required":["status"]},"UserProfileResponse":{"type":"object","description":"Authenticated user profile","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]}}},"UserDashboardStatsResponse":{"type":"object","description":"Complaint statistics for the authenticated user","properties":{"totalComplaints":{"type":"integer","format":"int64"},"pendingComplaints":{"type":"integer","format":"int64"},"resolvedComplaints":{"type":"integer","format":"int64"}}},"PageComplaintResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ComplaintResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"DashboardStatsResponse":{"type":"object","properties":{"totalComplaints":{"type":"integer","format":"int64"},"pendingComplaints":{"type":"integer","format":"int64"},"inProgressComplaints":{"type":"integer","format":"int64"},"resolvedComplaints":{"type":"integer","format":"int64"}}},"MonthlyComplaintAnalyticsResponse":{"type":"object","properties":{"year":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"complaintCount":{"type":"integer","format":"int64"}}},"CategoryAnalyticsResponse":{"type":"object","properties":{"category":{"type":"string"},"complaintCount":{"type":"integer","format":"int64"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}